# web/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: 'GLOBALPAS API-list'
  version: '1.0'
paths:
  /v1/author/:
    get:
      tags:
        - author
      operationId: get-all-authors
      responses:
        '200':
          description: get-all-books
  /v1/author/create:
    post:
      tags:
        - author
      operationId: create-new-author
      requestBody:
        description: 'Book data in form-data format'
        required: true
        content:
          multipart/form-data:
            schema:
              properties:
                name:
                  description: 'Name of the Author'
                  type: string
              type: object
      responses:
        '200':
          description: 'Create a new book'
  /v1/author/update:
    put:
      tags:
        - author
      summary: 'Update author information'
      operationId: updateAuthor
      parameters:
        -
          name: id
          in: query
          description: 'ID of the author to be updated'
          required: true
          schema:
            type: integer
      requestBody:
        description: 'Author data in form-data format'
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              properties:
                name:
                  description: 'Updated name of the author'
                  type: string
              type: object
      responses:
        '200':
          description: 'Author updated successfully'
        '404':
          description: 'Author not found'
  /v1/books/:
    get:
      tags:
        - books
      operationId: get-all-books
      responses:
        '200':
          description: get-all-books
  /v1/books/create:
    post:
      tags:
        - books
      operationId: create-new-books
      requestBody:
        description: 'Book data in form-data format'
        required: true
        content:
          multipart/form-data:
            schema:
              properties:
                title:
                  description: 'Title of the book'
                  type: string
                author_id:
                  description: 'ID of the author'
                  type: integer
                book_page:
                  description: 'Number of pages in the book'
                  type: integer
                language:
                  description: 'ID of the language'
                  type: integer
                genre:
                  description: 'ID of the genre'
                  type: integer
              type: object
      responses:
        '200':
          description: 'Create a new book'
  /v1/books/search-books:
    get:
      tags:
        - books
      operationId: SearchBooks
      parameters:
        -
          name: title
          in: query
          description: 'Book title'
          required: false
          schema:
            type: string
        -
          name: minPage
          in: query
          description: 'Minimum book page'
          required: false
          schema:
            type: integer
            maximum: 555
            minimum: 111
        -
          name: maxPage
          in: query
          description: 'Maximum book page'
          required: false
          schema:
            type: integer
            maximum: 555
            minimum: 111
        -
          name: 'author_id[]'
          in: query
          description: 'array of authors'
          required: false
          style: form
          schema:
            type: array
            items:
              type: integer
        -
          name: 'language_id[]'
          in: query
          description: 'array of languages'
          required: false
          style: form
          schema:
            type: array
            items:
              type: integer
        -
          name: genre
          in: query
          description: genre
          required: false
          schema:
            type: string
            enum:
              - 1
              - 2
              - 3
              - 4
              - 5
      responses:
        '200':
          description: ''